afsocket: move source driver machinery to lib/socket
