+++ lib/socket/Makefile.am
@@ -1,0 +1,20 @@
+socketincludedir			= ${pkgincludedir}/socket
+
+EXTRA_DIST += lib/socket/CMakeLists.txt
+
+socketinclude_HEADERS = \
+	lib/socket/afinet.h	\
+	lib/socket/afsocket-signals.h	\
+	lib/socket/afsocket-source.h	\
+	lib/socket/socket-options.h	\
+	lib/socket/socket-options-inet.h	\
+	lib/socket/transport-mapper.h	\
+	lib/socket/transport-mapper-inet.h
+
+socket_sources = \
+	lib/socket/afinet.c	\
+	lib/socket/afsocket-source.c	\
+	lib/socket/socket-options.c	\
+	lib/socket/socket-options-inet.c	\
+	lib/socket/transport-mapper.c	\
+	lib/socket/transport-mapper-inet.c
