+++ modules/afsocket/CMakeLists.txt
@@ -1,11 +1,8 @@
 set(AFSOCKET_SOURCES
     afsocket.c
     afsocket.h
-    afsocket-source.c
-    afsocket-source.h
     afsocket-dest.c
     afsocket-dest.h
-    afsocket-signals.h
     afinet-source.c
     afinet-source.h
     afinet-dest-failover.c
@@ -42,10 +39,8 @@ add_module(
   TARGET afsocket
   GRAMMAR afsocket-grammar
   INCLUDES ${ZLIB_INCLUDE_DIRS}
-           ${WRAP_INCLUDE_DIRS}
            ${Libsystemd_INCLUDE_DIRS}
-  DEPENDS ${WRAP_LIBRARIES}
-          ${ZLIB_LIBRARIES}
+  DEPENDS ${ZLIB_LIBRARIES}
           OpenSSL::SSL
           OpenSSL::Crypto
           libnet
