+++ modules/afsocket/Makefile.am
@@ -1,11 +1,8 @@
 modules_afsocket_libafsocket_la_SOURCES	=	\
 	modules/afsocket/afsocket.c			\
 	modules/afsocket/afsocket.h			\
-	modules/afsocket/afsocket-source.c		\
-	modules/afsocket/afsocket-source.h		\
 	modules/afsocket/afsocket-dest.c		\
 	modules/afsocket/afsocket-dest.h		\
-	modules/afsocket/afsocket-signals.h		\
 	modules/afsocket/afinet-source.c		\
 	modules/afsocket/afinet-source.h		\
 	modules/afsocket/afinet-dest.c			\
@@ -45,7 +42,7 @@ modules_afsocket_libafsocket_la_CFLAGS	=	\
 modules_afsocket_libafsocket_la_LIBADD	=	\
 	$(MODULE_DEPS_LIBS) $(OPENSSL_LIBS)	\
 	$(ZLIB_LIBS) $(LIBNET_LIBS)	\
-	$(LIBWRAP_LIBS) $(libsystemd_LIBS)
+	$(libsystemd_LIBS)
 EXTRA_modules_afsocket_libafsocket_la_DEPENDENCIES= 	\
 	$(MODULE_DEPS_LIBS)
 modules_afsocket_libafsocket_la_LDFLAGS	= 	\
