+++ modules/afsocket/afinet-source.h
@@ -25,7 +25,7 @@
 #define AFINET_SOURCE_H_INCLUDED
 
 #include "socket/afinet.h"
-#include "afsocket-source.h"
+#include "socket/afsocket-source.h"
 #include "transport/tls-context.h"
 
 typedef struct _AFInetSourceDriver
